requirements.txt
Flask>=2.3.0
Flask-Login>=0.6.3
Flask-Bcrypt>=1.0.1
Flask-WTF>=1.1.1
Flask-Mail>=0.9.1
email_validator>=2.0.0
psycopg2-binary>=2.9.9
requests>=2.31.0
beautifulsoup4>=4.12.0
celery>=5.3.0
kombu[sqlalchemy]>=5.3.0
redis>=5.0.0
spacy>=3.7.0
PyPDF2>=3.0.1
python-docx>=1.1.0
gunicorn>=21.2.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-flask>=1.2.0
pytest-mock>=3.10.0
flask-testing>=0.8.1
flake8
coverage>=7.0.0